+++ files/system/usr/libexec/secureblue/audit_secureblue.py
@@ -18,129 +18,28 @@ Auditing script for secureblue. See https://secureblue.dev/ for more info.
 
 import argparse
 import asyncio
-import enum
 import filecmp
 import glob
 import json
 import os
 import os.path
-import re
 import signal
 import sys
-import textwrap
 import traceback
 
 # All subprocess calls we make have trusted inputs and do not use shell=True.
 import subprocess  # nosec
 
-from collections.abc import Iterable
-from typing import Final, Generator
-
-import rpm
-
-from auditor import AuditError, Report, Status, audit, bold, categorize, depends_on, global_audit
+from auditor import Report, Status, audit, bold, categorize, depends_on, global_audit
 from audit_flatpak import check_flatpak_permissions, parse_flatpak_permissions
-
-PASS: Final = Status.PASS
-INFO: Final = Status.INFO
-WARN: Final = Status.WARN
-FAIL: Final = Status.FAIL
-UNKNOWN: Final = Status.UNKNOWN
-
-
-def command_stdout(*args: str, check: bool = True) -> str:
-    """Run a command in the shell and return the contents of stdout."""
-    # We only call this with trusted inputs and do not set shell=True.
-    return subprocess.run(args, capture_output=True, check=check, text=True).stdout.strip()  # nosec
-
-
-class AsyncProcessError(AuditError):
-    """An asynchronous subprocess command returned a nonzero exit code."""
-
-
-async def async_command_stdout(cmd: str, *args: str, check: bool = True) -> str:
-    """Asynchronously run a command in the shell and return the contents of stdout."""
-    sub = await asyncio.create_subprocess_exec(
-        cmd, *args, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL
-    )
-    await sub.wait()
-    if check and sub.returncode != 0:
-        err = f"async command `{cmd} {' '.join(args)}` returned nonzero exit code {sub.returncode}"
-        raise AsyncProcessError(err)
-    if sub.stdout is None:
-        err = f"Failed to get stdout for async command `{cmd} {' '.join(args)}`"
-        raise AsyncProcessError(err)
-    output = await sub.stdout.read()
-    return output.decode("utf-8", errors="replace").strip()
-
-
-def command_succeeds(*args: str) -> bool:
-    """Run a command in the shell and return the contents of stdout."""
-    # We only call this with trusted inputs and do not set shell=True.
-    return subprocess.run(args, capture_output=True, check=False).returncode == 0  # nosec
-
-
-def parse_config(
-    stream: Iterable[str], *, sep: str = "=", comment: str = "#"
-) -> Generator[tuple[str, str | None]]:
-    """
-    Parse a text stream as a simple configuration file, yielding a sequence of keys and values
-    separated by the given separator ("=" by default).
-    """
-    for line in stream:
-        line = line.strip()
-        if not line or line.startswith(comment):
-            continue
-        split = line.split(sep, maxsplit=1)
-        key = split[0].strip()
-        if len(split) == 2:
-            value = split[1].strip()
-        else:
-            value = None
-        yield key, value
-
-
-def is_rpm_package_installed(name: str) -> bool:
-    """Checks if the given RPM package is installed."""
-    ts = rpm.TransactionSet()
-    matches = ts.dbMatch("name", name)
-    return len(matches) > 0
-
-
-class Image(enum.Enum):
-    """Fedora atomic base image"""
-
-    SILVERBLUE = enum.auto()
-    KINOITE = enum.auto()
-    SERICEA = enum.auto()
-    COSMIC = enum.auto()
-    COREOS = enum.auto()
-
-    @classmethod
-    def from_image_ref(cls, image_ref: str):
-        """Convert an image reference to the corresponding Image enum instance."""
-        if "silverblue" in image_ref:
-            return cls.SILVERBLUE
-        if "kinoite" in image_ref:
-            return cls.KINOITE
-        if "sericea" in image_ref:
-            return cls.SERICEA
-        if "cosmic" in image_ref:
-            return cls.COSMIC
-        if "securecore" in image_ref:
-            return cls.COREOS
-        return None
-
-
-###############################################################################
-# Checks to be run go below this line.
-###############################################################################
+from utils import print_err, warn_if_root, get_width, get_legend, handle_sigint, parse_config, command_stdout, Image, \
+    command_succeeds, validate_sysctl, get_flatpak_permissions
 
 
 @audit
 def audit_kargs():
     """Check for hardened kernel arguments."""
-    status = PASS
+    status = Status.PASS
     warnings = []
     rec = None
 
@@ -171,17 +70,17 @@ def audit_kargs():
     )
     for karg in kargs_expected:
         if karg not in kargs_current:
-            status = status.downgrade_to(FAIL)
+            status = status.downgrade_to(Status.FAIL)
             warnings.append(f"Missing kernel argument: {karg}")
 
     karg_32bit = "ia32_emulation=0"
     if karg_32bit not in kargs_current:
-        status = status.downgrade_to(WARN)
+        status = status.downgrade_to(Status.WARN)
         warnings.append(f"Missing kernel argument: {karg_32bit} (32-bit support)")
 
     karg_nosmt = "nosmt=force"
     if karg_nosmt not in kargs_current:
-        status = status.downgrade_to(WARN)
+        status = status.downgrade_to(Status.WARN)
         warnings.append(f"Missing kernel argument: {karg_nosmt} (force-disable SMT)")
 
     kargs_expected_unstable = (
@@ -193,27 +92,14 @@ def audit_kargs():
     )
     for karg in kargs_expected_unstable:
         if karg not in kargs_current:
-            status = status.downgrade_to(WARN)
+            status = status.downgrade_to(Status.WARN)
             warnings.append(f"Missing kernel argument (unstable): {karg}")
 
-    if status != PASS:
+    if status != Status.PASS:
         rec = """To set hardened kernel arguments, run:
             $ ujust set-kargs-hardening"""
 
     yield Report("Checking for hardened kernel arguments", status, warnings=warnings, recs=rec)
-
-
-def validate_sysctl(sysctl: str, actual: str, expected: str) -> bool:
-    """Validate a sysctl value against an expected value."""
-    actual = re.sub(r"\s+", " ", actual.strip())
-    replace = {"disabled": "0", "enabled": "1"}.get(actual)
-    if replace is not None:
-        actual = replace
-    if sysctl == "kernel.sysrq":
-        # Both 0 and 4 are secure values for this setting. For details, see:
-        # https://www.kernel.org/doc/html/latest/admin-guide/sysrq.html
-        return actual in (expected, "0", "4")
-    return actual == expected
 
 
 @audit
@@ -224,12 +110,12 @@ def audit_sysctl():
     sysctl_expected = {}
     for key, value in parse_config(conf):
         sysctl_expected[key] = value
-    status = PASS
+    status = Status.PASS
     sysctl_errors = []
     with open("/etc/sysctl.d/60-hardening.conf", "r", encoding="utf-8") as f:
         etc_conf = f.readlines()
     if conf != etc_conf:
-        status = WARN
+        status = Status.WARN
         sysctl_errors.append("/etc/sysctl.d/60-hardening.conf has been modified")
     for sysctl, expected in sysctl_expected.items():
         sysctl_path = f"/proc/sys/{sysctl.replace('.', '/')}"
@@ -240,7 +126,7 @@ def audit_sysctl():
             except PermissionError:
                 continue
             if not validate_sysctl(sysctl, actual, expected):
-                status = FAIL
+                status = Status.FAIL
                 sysctl_errors.append(f"{sysctl} should be {expected}, found {actual}")
                 break
     yield Report("Ensuring no sysctl overrides", status, warnings=sysctl_errors)
@@ -253,10 +139,10 @@ def audit_signed_image(state):
     image_ref = json.loads(ostree_status)["deployments"][0]["container-image-reference"]
     state["image"] = Image.from_image_ref(image_ref)
     if image_ref.startswith("ostree-image-signed:"):
-        status = PASS
+        status = Status.PASS
         recs = None
     else:
-        status = FAIL
+        status = Status.FAIL
         recs = """The current image is not signed.
             To rebase to a signed image, download and run or re-run install_secureblue.sh
             from the secureblue GitHub repository."""
@@ -280,14 +166,14 @@ def audit_modprobe(state):
             if mod in blacklisted_modules:
                 unwanted_modules.append(mod)
     unwanted_modules.sort()
-    status = PASS
+    status = Status.PASS
     warnings = []
     with open("/etc/modprobe.d/blacklist.conf", "r", encoding="utf-8") as f:
         if f.readlines() != conf:
-            status = WARN
+            status = Status.WARN
             warnings.append("/etc/modprobe.d/blacklist.conf has been modified")
     for mod in unwanted_modules:
-        status = FAIL
+        status = Status.FAIL
         warnings.append(f"{mod} is in blacklist.conf but it is loaded")
     state["bluetooth_loaded"] = "bluetooth" in unwanted_modules
     yield Report("Ensuring no modprobe overrides", status, warnings=warnings)
@@ -301,10 +187,10 @@ def audit_ptrace(state):
     state["ptrace_allowed"] = ptrace_scope < 3
     match ptrace_scope:
         case 3:
-            status = PASS
+            status = Status.PASS
             rec = None
         case 0:
-            status = FAIL
+            status = Status.FAIL
             rec = f"""ptrace is allowed and {bold("unrestricted")} (ptrace_scope = 0)!
                 For more info on what this means, see:
                 https://www.kernel.org/doc/html/latest/admin-guide/LSM/Yama.html
@@ -312,7 +198,7 @@ def audit_ptrace(state):
                 $ ujust toggle-ptrace-scope
                 To allow restricted ptrace, run the above command twice."""
         case _:
-            status = WARN
+            status = Status.WARN
             rec = f"""ptrace is allowed, but restricted (ptrace_scope = {ptrace_scope}).
                 For more info on what this means, see:
                 https://www.kernel.org/doc/html/latest/admin-guide/LSM/Yama.html
@@ -326,24 +212,24 @@ def audit_authselect():
     """Ensure no authselect overrides have been made."""
     cmp = filecmp.dircmp("/usr/etc/authselect", "/etc/authselect", shallow=False, ignore=[])
     if cmp.left_only or cmp.right_only or cmp.diff_files or cmp.funny_files:
-        status = FAIL
+        status = Status.FAIL
     else:
-        status = PASS
+        status = Status.PASS
     yield Report("Ensuring no authselect overrides", status)
 
 
 @audit
 def audit_container_policy():
     """Ensure container policy has not been modified."""
-    status = PASS
+    status = Status.PASS
     warnings = []
     policy_file = "/etc/containers/policy.json"
     if not filecmp.cmp(f"/usr{policy_file}", policy_file):
-        status = FAIL
+        status = Status.FAIL
         warnings.append(f"{policy_file} has been modified")
     local_override = "~/.config/containers/policy.json"
     if os.path.isfile(os.path.expanduser(local_override)):
-        status = FAIL
+        status = Status.FAIL
         warnings.append(f"{local_override} exists")
     yield Report("Ensuring no container policy overrides", status, warnings=warnings)
 
@@ -352,10 +238,10 @@ def audit_container_policy():
 def audit_unconfined_userns():
     """Ensure unconfined-domain processes cannot create user namespaces."""
     if command_stdout("ujust", "check-unconfined-userns-state") == "disabled":
-        status = PASS
+        status = Status.PASS
         recs = None
     else:
-        status = FAIL
+        status = Status.FAIL
         recs = """Unconfined domain user namespace creation is permitted.
                 To disallow it, run:
                 $ ujust toggle-unconfined-domain-userns-creation"""
@@ -366,10 +252,10 @@ def audit_unconfined_userns():
 def audit_container_userns():
     """Ensure container-domain processes cannot create user namespaces."""
     if command_stdout("ujust", "check-container-userns-state") == "disabled":
-        status = PASS
+        status = Status.PASS
         recs = []
     else:
-        status = WARN
+        status = Status.WARN
         recs = """Container domain user namespace creation is permitted.
                 To disallow it, run:
                 $ ujust toggle-container-domain-userns-creation"""
@@ -380,10 +266,10 @@ def audit_container_userns():
 def audit_usbguard():
     """Ensure usbguard is active."""
     if command_succeeds(*"systemctl is-active --quiet usbguard".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """USBGuard is not active. To set up USBGuard, run:
             $ ujust setup-usbguard
             Caution: if you have already set up USBGuard, this will overwrite the
@@ -395,10 +281,10 @@ def audit_usbguard():
 def audit_chronyd():
     """Ensure chronyd is active."""
     if command_succeeds(*"systemctl is-active --quiet chronyd".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """chronyd is inactive.
             To start and enable it, run:
             $ systemctl enable --now chronyd"""
@@ -422,25 +308,25 @@ def audit_dns():
                     elif key == "DNSOverTLS":
                         dot = value
         except FileNotFoundError:
-            status = FAIL
+            status = Status.FAIL
         except PermissionError:
-            status = UNKNOWN
+            status = Status.UNKNOWN
             warning = f"Unable to read file {conf_path}"
         else:
             if dnssec == "true" and dot == "true":
-                status = PASS
+                status = Status.PASS
             elif dot == "opportunistic":
-                status = WARN
+                status = Status.WARN
             else:
-                status = FAIL
-        if status in (WARN, FAIL):
+                status = Status.FAIL
+        if status in (Status.WARN, Status.FAIL):
             caveat = " (opportunistic DNS-over-TLS only)" if dot == "opportunistic" else ""
             rec = f"""System DNS resolution is not secure{caveat}.
                     To select a secure resolver, run:
                     $ ujust dns-selector
                     If you are using a VPN, you may want to disregard this recommendation."""
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """systemd-resolved is inactive.
                 To start and enable it, run:
                 $ systemctl enable --now systemd-resolved"""
@@ -450,7 +336,7 @@ def audit_dns():
 @audit
 def audit_mac_randomization():
     """Ensure MAC randomization is enabled."""
-    status = FAIL
+    status = Status.FAIL
     warning = None
     conf_path = "/etc/NetworkManager/conf.d/rand_mac.conf"
     try:
@@ -463,14 +349,14 @@ def audit_mac_randomization():
                 if key == "wifi.cloned-mac-address" and value in ["random", "stable"]:
                     wifi = True
                 if ethernet and wifi:
-                    status = PASS
+                    status = Status.PASS
                     break
     except FileNotFoundError:
         pass
     except PermissionError:
-        status = UNKNOWN
+        status = Status.UNKNOWN
         warning = f"Unable to read file {conf_path}"
-    if status == FAIL:
+    if status == Status.FAIL:
         rec = """MAC randomization is not enabled.
                 To enable it, run:
                 $ ujust toggle-mac-randomization"""
@@ -483,10 +369,10 @@ def audit_mac_randomization():
 def audit_rpm_ostree_timer():
     """Ensure rpm-ostree automatic updates are enabled."""
     if command_succeeds(*"systemctl is-enabled --quiet rpm-ostreed-automatic.timer".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """rpm-ostreed-automatic.timer is disabled.
                 To enable, run:
                 $ systemctl enable --now rpm-ostreed-automatic.timer"""
@@ -497,10 +383,10 @@ def audit_rpm_ostree_timer():
 def audit_podman_auto_update():
     """Ensure podman automatic updates are enabled."""
     if command_succeeds(*"systemctl is-enabled --quiet podman-auto-update.timer".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """podman-auto-update.timer is disabled.
                 To enable, run:
                 $ systemctl enable --now podman-auto-update.timer"""
@@ -511,10 +397,10 @@ def audit_podman_auto_update():
 def audit_podman_global_auto_update():
     """Ensure podman automatic updates are enabled globally."""
     if command_succeeds(*"systemctl --global is-enabled --quiet podman-auto-update.timer".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """podman-auto-update.timer is not enabled globally.
                 To enable, run:
                 $ systemctl enable --global podman-auto-update.timer"""
@@ -527,20 +413,20 @@ def audit_flatpak_auto_update():
     if not command_succeeds(*"command -v flatpak".split()):
         return
     if command_succeeds(*"systemctl --global is-enabled --quiet flatpak-user-update.timer".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """flatpak-user-update.timer is not enabled globally.
                 To enable, run:
                 $ systemctl enable --global flatpak-user-update.timer"""
     yield Report("Ensuring flatpak-user-update.timer is enabled globally", status, recs=rec)
 
     if command_succeeds(*"systemctl is-enabled --quiet flatpak-system-update.timer".split()):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """flatpak-system-update.timer is not enabled globally.
                 To enable, run:
                 $ systemctl enable --now flatpak-system-update.timer"""
@@ -554,10 +440,10 @@ def audit_wheel():
         rec = f"""Current user is in the wheel group.
             To set up a separate wheel account, follow the instructions here:
             {bold("https://secureblue.dev/install#wheel")}"""
-        status = FAIL
+        status = Status.FAIL
     else:
         rec = None
-        status = PASS
+        status = Status.PASS
     yield Report("Ensuring user is not a member of wheel", status, recs=rec)
 
 
@@ -578,10 +464,10 @@ def audit_xwayland(state):
         case _:
             return
     if os.path.isfile(path):
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = f"""Xwayland is enabled for {de}. To disable, run:
             $ ujust toggle-xwayland"""
     yield Report(f"Ensuring xwayland is disabled for {de}", status, recs=rec)
@@ -597,10 +483,10 @@ def audit_gnome_extensions(state):
         *"command -p gsettings get org.gnome.shell allow-extension-installation".split()
     )
     if allowed == "false":
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """GNOME user extensions are enabled. To disable, run:
             $ ujust toggle-gnome-extensions"""
     yield Report("Ensuring GNOME user extensions are disabled", status, recs=rec)
@@ -610,10 +496,10 @@ def audit_gnome_extensions(state):
 def audit_selinux():
     """Ensure SELinux is in enforcing mode."""
     if command_stdout("getenforce") == "Enforcing":
-        status = PASS
+        status = Status.PASS
         rec = None
     else:
-        status = FAIL
+        status = Status.FAIL
         rec = """SELinux is in Permissive mode.
             To set to Enforcing mode, run:
             $ run0 setenforce 1"""
@@ -625,16 +511,16 @@ def audit_environment_file():
     """Ensure /etc/environment has not been modified."""
     try:
         if filecmp.cmp("/usr/etc/environment", "/etc/environment"):
-            status = PASS
+            status = Status.PASS
             warning = None
         else:
-            status = WARN
+            status = Status.WARN
             warning = "/etc/environment has been modified"
     except FileNotFoundError:
-        status = WARN
+        status = Status.WARN
         warning = "/etc/environment has been deleted"
     except PermissionError:
-        status = WARN
+        status = Status.WARN
         warning = "/etc/environment cannot be read"
     yield Report("Ensuring no environment file overrides", status, warnings=warning)
 
@@ -645,18 +531,18 @@ def audit_kde_ghns(state):
     """Ensure KDE GHNS is disabled."""
     if state["image"] != Image.KINOITE:
         return
-    status = FAIL
+    status = Status.FAIL
     warning = None
     try:
         with open("/etc/xdg/kdeglobals", "r", encoding="utf-8") as f:
             for key, value in parse_config(f):
                 if key == "ghns" and value == "false":
-                    status = PASS
+                    status = Status.PASS
                     break
     except (FileNotFoundError, PermissionError):
-        status = WARN
+        status = Status.WARN
         warning = "/etc/xdg/kdeglobals not found or inaccessible"
-    if status == FAIL:
+    if status == Status.FAIL:
         rec = """KDE GHNS is enabled.
             To disable, run:
             $ ujust toggle-ghns"""
@@ -673,27 +559,27 @@ def audit_hardened_malloc():
         with open("/etc/ld.so.preload", "r", encoding="utf-8") as f:
             preloaded = f.read().split()
     except FileNotFoundError:
-        status = FAIL
+        status = Status.FAIL
         warnings.append("ld.so.preload not found")
     except PermissionError:
-        status = FAIL
+        status = Status.FAIL
         warnings.append("Permission denied to read ld.so.preload")
     else:
         if preloaded == ["libhardened_malloc.so"]:
-            status = PASS
+            status = Status.PASS
         elif "libhardened_malloc.so" in preloaded:
-            status = WARN
+            status = Status.WARN
             warnings.append("hardened_malloc set, but ld.so.preload has been modified")
         elif "libhardened_malloc-light.so" in preloaded:
-            status = WARN
+            status = Status.WARN
             warnings.append("'light' variant of hardened_malloc set")
         elif "libhardened_malloc-pkey.so" in preloaded:
-            status = WARN
+            status = Status.WARN
             warnings.append("'pkey' variant of hardened_malloc set")
         else:
-            status = FAIL
+            status = Status.FAIL
             warnings.append("hardened_malloc not set")
-    if status == PASS:
+    if status == Status.PASS:
         rec = None
     else:
         rec = """/etc/ld.so.preload has been modified.
@@ -708,9 +594,9 @@ def audit_hardened_malloc():
 def audit_secureboot():
     """Ensure secureboot is enabled."""
     if command_stdout("mokutil", "--sb-state", check=False) == "SecureBoot enabled":
-        status = PASS
+        status = Status.PASS
     else:
-        status = FAIL
+        status = Status.FAIL
     yield Report("Ensuring secure boot is enabled", status)
 
 
@@ -744,14 +630,15 @@ def audit_bash_env_lockdown():
             if not immutable:
                 unlocked_files.append(path)
     if unlocked_files:
-        status = FAIL
+        status = Status.FAIL
+        unlocked_files_string="\n".join(unlocked_files)
         rec = f"""Bash environment is not locked down.
                 The following files do not appear to be immutable or do not exist:
-                {"\n".join(unlocked_files)}
+                {unlocked_files_string}
                 To fix, run:
                 $ ujust toggle-bash-environment-lockdown"""
     else:
-        status = PASS
+        status = Status.PASS
         rec = None
     yield Report("Ensuring current user's bash environment is locked down", status, recs=rec)
 
@@ -773,20 +660,14 @@ def audit_flatpak_remotes():
             "https://dl.flathub.org/repo/",
             "https://dl.flathub.org/beta-repo/",
         ]:
-            status = FAIL
+            status = Status.FAIL
             warnings.append(f"{name} is configured with an unknown url")
         elif subset != "verified":
-            status = FAIL
+            status = Status.FAIL
             warnings.append(f"{name} is not a verified repo")
         else:
-            status = PASS
+            status = Status.PASS
         yield Report(f"Auditing flatpak remote {name}", status, warnings=warnings)
-
-
-async def get_flatpak_permissions(name: str, version: str) -> str:
-    """Get permissions for an installed flatpak."""
-    return await async_command_stdout("flatpak", "info", "--show-permissions", name, version)
-
 
 @audit
 @categorize("flatpak")
@@ -814,99 +695,23 @@ async def audit_flatpak_permissions(state):
     for name, version in flatpaks:
         perms_text = await tasks[(name, version)]
         perms = parse_flatpak_permissions(perms_text)
-        status, warnings, recs = check_flatpak_permissions(
+        flatpak_permissions_state = check_flatpak_permissions(
             name, perms, state["bluetooth_loaded"], state["ptrace_allowed"]
         )
         if version == "stable":
             report_text = f"Auditing {name}"
         else:
             report_text = f"Auditing {name} ({version})"
-        yield Report(report_text, status, warnings=warnings, recs=recs)
-
+        yield Report(
+            report_text,
+            flatpak_permissions_state.status,
+            warnings=flatpak_permissions_state.warnings,
+            recs=flatpak_permissions_state.recs
+        )
 
 ###############################################################################
 # Checks to be run go above this line.
 ###############################################################################
-
-
-def print_err(text: str):
-    """Print text to stderr in bold and red."""
-    print(f"\x1b[1m\x1b[31m{text}\x1b[0m", file=sys.stderr)
-
-
-def handle_sigint(_sig, _frame):
-    """Gracefully handle interrupt signal."""
-    print_err("\n[Audit process interrupted. Exiting.]")
-    # Suppress output from exceptions in unfinished tasks
-    sys.stderr = None
-    sys.exit(1)
-
-
-def warn_if_root():
-    """If run as root, warn that this is not recommended."""
-    if os.getuid() == 0:
-        print_err("\n*** WARNING: Running audit script as root is not recommended. ***")
-        print_err("*** Some results may be misleading or incomplete. ***\n")
-
-
-def get_width() -> int:
-    """Get the width in columns to be used for reports."""
-    try:
-        width = min(max(80, os.get_terminal_size().columns), 100)
-    except OSError:
-        width = 80
-    return width
-
-
-def format_legend_entry(status: Status, description: str, width: int = 80) -> str:
-    """Format legend entry"""
-    key_str = f"[{status.to_str_in_color()}]: "
-    key_str_width = len(status.name) + 4
-    description = re.sub(r"\s+", " ", description.strip())
-    lines = textwrap.wrap(description, width=width - key_str_width)
-    if not lines:
-        return f"{key_str}\n"
-    entry = f"{key_str}{lines[0]}\n"
-    for line in lines[1:]:
-        entry += f"{' ' * key_str_width}{line}\n"
-    return entry
-
-
-def get_legend(width: int = 80) -> str:
-    """Get legend to be printed with --help option."""
-    legend = "The following status indicators accompany checks run by the audit script:\n\n"
-    status_descriptions: dict[Status, str] = {
-        FAIL: "check failed - the configuration may be less secure.",
-        WARN: "partial failure, or less significant issue detected.",
-        PASS: "check passed - no problems detected.",
-        UNKNOWN: "unable to perform check (usually due to a file permission issue).",
-    }
-    for status, desc in status_descriptions.items():
-        legend += format_legend_entry(status, desc, width)
-    legend += "\nFor flatpak checks, the status indicators have more specific meanings:\n\n"
-    flatpak_status_descriptions: dict[Status, str] = {
-        FAIL: """app has permissions that can be used as sandbox escapes, allow it to modify
-            its own permissions, or otherwise grant very broad access to the system (e.g. access
-            to certain directories, direct D-Bus access, X11).""",
-        WARN: """app has permissions that have some sandbox escape potential or otherwise
-            weaken security (e.g. PulseAudio, Bluetooth, not using hardened_malloc).""",
-        INFO: """no potential sandbox escapes detected but some permissions could increase
-            attack surface or have privacy implications (e.g. network access).""",
-        PASS: "no app permissions flagged (however, not all permissions are audited).",
-    }
-    for status, desc in flatpak_status_descriptions.items():
-        legend += format_legend_entry(status, desc, width)
-    legend += "\n" + textwrap.fill(
-        textwrap.dedent(
-            """\
-            Note that some flatpak apps require broad permissions to function. Permissions being
-            flagged by the audit script do not necessarily mean that action should be taken.
-            """
-        ),
-        width=width,
-    )
-    return legend
-
 
 async def main() -> int:
     """Main entry point. Parse command-line arguments and run audit."""
@@ -943,7 +748,6 @@ async def main() -> int:
         print_err("\n*** WARNING: Unexpected error occurred. ***")
         return 1
     return 0
-
 
 if __name__ == "__main__":
     sys.exit(asyncio.run(main()))
